Refactoring and index loading from disk
