+++ data/ibf_example.fna
@@ -1,0 +1,6 @@
+>Sequence 1
+ACTGACTGACTGATC
+>Sequence 2
+GTGACTGACTGACTCG
+>Sequence 3
+AAAAAAACGATCGACA
