+++ include/query.h
@@ -1,0 +1,8 @@
+//
+// Created by Remy Schwab on 20.09.22.
+//
+
+#ifndef KBIOREG_QUERY_H
+#define KBIOREG_QUERY_H
+
+#endif //KBIOREG_QUERY_H
