+++ src/CMakeLists.txt
@@ -8,7 +8,7 @@ target_include_directories ("${PROJECT_NAME}_interface" INTERFACE ../include)
 target_compile_options ("${PROJECT_NAME}_interface" INTERFACE "-pedantic" "-Wall" "-Wextra")
 
 # An object library (without main) to be used in multiple targets.
-add_library ("${PROJECT_NAME}_lib" STATIC nfa_pointer.cpp graphMaker.cpp korotkov_nfa.cpp utils.cpp index.cpp)
+add_library ("${PROJECT_NAME}_lib" STATIC nfa_pointer.cpp graphMaker.cpp korotkov_nfa.cpp utils.cpp index.cpp query.cpp)
 target_link_libraries ("${PROJECT_NAME}_lib" PUBLIC "${PROJECT_NAME}_interface")
 
 add_executable ("${PROJECT_NAME}" main.cpp)
